==> 3H_obtain_robust_markers.ipynb <==
# Obtaining robust cell population markers, and redefining/reassuring the biased cell populations

## imports

%load_ext autoreload
%autoreload 2

import scanpy as sc
import scanpy.external as sce
import pandas as pd
import numpy as np
import os
import triku as tk
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib as mpl
from tqdm.notebook import tqdm
import scipy.sparse as spr
import matplotlib.cm as cm
import networkx as nx
import pylab

!pip install cellassign --upgrade

# local imports and imports from other notebooks
from cellassign import assign_cats
from fb_functions import make_gene_scoring_with_expr, plot_score_graph, plot_UMAPS_gene, plot_adata_cluster_properties
%store -r dict_colors_human
%store -r seed
%store -r magma
%store -r data_dir

%store -r plot_params

pylab.rcParams.update(plot_params)
pd.set_option('display.max_columns', None)
pd.options.display.float_format = "{:,.2f}".format

plot_params

**IMPORTANT: I am running this analysis in a computer with ~500 GB of RAM. I will load many datasets at once, which might be too much for some computers. I took this decision conciously, to have as much info available at any time as possible. If you cannot run all the analysis at once, you can run it by parts.**

## Anndata loading

ahlers_2022_dir = data_dir + '/ahlers_2022'
ahlers_2022_young_human_fb = sc.read(ahlers_2022_dir + '/ahlers_2022_young_human_fb_processed.h5')

billi_2022_dir = data_dir + '/billi_2022'
billi_2022_ctrl_human_fb = sc.read(billi_2022_dir + '/billi_2022_ctrl_human_fb_processed.h5')

boothby_2021_dir = data_dir + '/boothby_2021'
boothby_2021_ctrl_human_fb = sc.read(boothby_2021_dir + '/boothby_2021_ctrl_human_fb_processed.h5')

burja_2022_dir = data_dir + '/burja_2022'
burja_2022_ctrl_human_fb = sc.read(burja_2022_dir + '/burja_2022_ctrl_human_fb_processed.h5')

deng_2021_dir = data_dir + '/deng_2021'
deng_2021_scar_human_fb = sc.read(deng_2021_dir + '/deng_2021_scar_human_fb_processed.h5')

gao_2021_dir = data_dir + '/gao_2021'
gao_2021_ctrl_human_fb = sc.read(gao_2021_dir + '/gao_2021_ctrl_human_fb_processed.h5')

gaydosik_2020_dir = data_dir + '/gaydosik_2020'
gaydosik_2020_ctrl_human_fb = sc.read(gaydosik_2020_dir + '/gaydosik_2020_ctrl_human_fb_processed.h5')

gur_2022_dir = data_dir + '/gur_2022'
gur_2022_ctrl_human_fb = sc.read(gur_2022_dir + '/gur_2022_ctrl_human_fb_processed.h5')

he_2020_dir = data_dir + '/He_2020'
he_2020_ctrl_human_fb = sc.read(he_2020_dir + '/he_2020_ctrl_human_fb_processed.h5')

hughes_2020_dir = data_dir + '/hughes_2020'
hughes_2020_ctrl_human_fb = sc.read(hughes_2020_dir + '/hughes_2020_ctrl_human_fb_processed.h5')

kim_2020_dir = data_dir + '/Kim_2020'
kim_2020_ctrl_human_fb = sc.read(kim_2020_dir + '/kim_2020_ctrl_human_fb_processed.h5')

liu_2021_dir = data_dir + '/liu_2021'
liu_2021_ctrl_human_fb = sc.read(liu_2021_dir + '/liu_2021_ctrl_human_fb_processed.h5')

mariottoni_2021_dir = data_dir + '/mariottoni_2021'
mariottoni_2021_ctrl_human_fb = sc.read(mariottoni_2021_dir + '/mariottoni_2021_ctrl_human_fb_processed.h5')

mirizio_2020_dir = data_dir + '/mirizio_2020'
mirizio_2020_scl_human_fb = sc.read(mirizio_2020_dir + '/mirizio_2020_scl_human_fb_processed.h5')

reynolds_2021_dir = data_dir + '/reynolds_2021'
reynolds_2021_ctrl_human_fb = sc.read(reynolds_2021_dir + '/reynolds_2021_ctrl_human_fb_processed.h5')

rindler_2021_dir = data_dir + '/rindler_2021'
rindler_2021_ctrl_human_fb = sc.read(rindler_2021_dir + '/rindler_2021_ctrl_human_fb_processed.h5')

sole_2020_dir = data_dir + '/Sole-Boldo_2020'
sole_2020_young_human_fb = sc.read(sole_2020_dir + '/sole_2020_young_human_fb_processed.h5')

tabib_2018_dir = data_dir + '/Tabib_2018'
tabib_2018_ctrl_human_fb = sc.read(tabib_2018_dir + '/tabib_2018_ctrl_human_fb_processed.h5')

tsc_2021_dir = data_dir + '/Tabula_Sapiens_Consortium_2021'
tsc_2021_ctrl_human_fb = sc.read(tsc_2021_dir + '/tsc_2021_ctrl_human_fb_processed.h5')

tabib_2021_dir = data_dir + '/Tabib_2021'
tabib_2021_ctrl_human_fb = sc.read(tabib_2021_dir + '/tabib_2021_ctrl_human_fb_processed.h5')

theo_2020_dir = data_dir + '/Theocharidis_2020/'
theo_2020_ctrl_human_dm_fb = sc.read(theo_2020_dir + '/theo_2020_ctrl_human_dm_fb_processed.h5')

theo_2021_dir = data_dir + '/Theocharidis_2021/'
theo_2021_ctrl_human_fb = sc.read(theo_2021_dir + '/theo_2021_ctrl_human_fb_processed.h5')

vors_2020_dir = data_dir + '/Vorstandlechner_2020'
vors_2020_ctrl_human_fb = sc.read(vors_2020_dir + '/vors_2020_ctrl_human_fb_processed.h5')

vors_2021_dir = data_dir + '/Vorstandlechner_2021'
vors_2021_ctrl_human_fb = sc.read(vors_2021_dir + '/vors_2021_ctrl_human_fb_processed.h5')

xu_2021_dir = data_dir + '/xu_2021'
xu_2021_ctrl_human_fb = sc.read(xu_2021_dir + '/xu_2021_ctrl_human_fb_processed.h5')

## Evaluate marker score for adatas
All the **commented anndatas are discarded** for marker selection because they do not show enough marker consistency and might bias the result, or because they belong to dataset from non control/healthy conditions.

list_datasets_human = [
                 ahlers_2022_young_human_fb,
                 billi_2022_ctrl_human_fb,
#                  boothby_2021_ctrl_human_fb,
#                  burja_2022_ctrl_human_fb,
                 deng_2021_scar_human_fb, 
                 gao_2021_ctrl_human_fb, 
                 gaydosik_2020_ctrl_human_fb, 
#                  gur_2022_ctrl_human_fb, 
#                  he_2020_ctrl_human_fb, 
#                  hughes_2020_ctrl_human_fb, 
                 kim_2020_ctrl_human_fb, 
                 liu_2021_ctrl_human_fb, 
#                  mariottoni_2021_ctrl_human_fb,
#                  mirizio_2020_scl_human_fb, 
#                  reynolds_2021_ctrl_human_fb,
                 rindler_2021_ctrl_human_fb,
                 sole_2020_young_human_fb, 
                 tabib_2018_ctrl_human_fb, 
                 tabib_2021_ctrl_human_fb,
#                  theo_2020_ctrl_human_dm_fb, 
#                  theo_2021_ctrl_human_fb,
#                  tsc_2021_ctrl_human_fb,
                 vors_2020_ctrl_human_fb, 
#                  vors_2021_ctrl_human_fb,
#                  xu_2021_ctrl_human_fb
                    ]

%store list_datasets_human

list_accepted_clusters_human = ['A1', 'A2', 'A3', 'A4', 'B1', 'B2', 'B3', 'B4', 'C1', 'C2', 'C3', 'C5', 'D1', 'D2', 'E1', 'T1',]                
list_accepted_axis_human = ['A', 'B', 'C', 'D', 'E']

# manual_axis is to create a robust set of markers
for adata in list_datasets_human:
    adata.obs['manual_axis'] = [i[0] for i in adata.obs['cluster']]

dict_make_gene_scoring = make_gene_scoring_with_expr(list_datasets=list_datasets_human, calculate_DEGs = True, group_name = 'cluster', value_ref = 'scores', select_method = 'pval', 
                                                     select_thres = 0.05, list_clusters=list_accepted_clusters_human)

dict_make_gene_scoring_axis = make_gene_scoring_with_expr(list_datasets=list_datasets_human, calculate_DEGs = True, group_name = 'manual_axis', list_clusters=['A', 'B', 'C', 'D', 'E'],
                                                     value_ref = 'scores', select_method = 'pval', select_thres = 0.05)

## Recalculating clusters in the datasets

dict_cats_clusters_robust = {i: np.array(dict_make_gene_scoring[i].index[:30]) for i in dict_make_gene_scoring.keys()}
dict_cats_axes_robust = {i: np.array(dict_make_gene_scoring_axis[i].index[:30]) for i in dict_make_gene_scoring_axis.keys()}

dict_cats_clusters_robust_3H = dict_cats_clusters_robust
%store dict_cats_clusters_robust_3H 

dict_cats_clusters_robust

dict_cats_axes_robust

### Ahlers 2022

assign_cats(ahlers_2022_young_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.8, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(ahlers_2022_young_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

ahlers_2022_young_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(ahlers_2022_young_human_fb.obs['cluster_robust']))]
sc.pl.umap(ahlers_2022_young_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(ahlers_2022_young_human_fb, cluster_column='cluster_robust')
ahlers_2022_young_human_fb.write_h5ad(ahlers_2022_dir + '/ahlers_2022_young_human_fb_robust.h5')

### Billy 2022

assign_cats(billi_2022_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.95, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(billi_2022_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

billi_2022_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(billi_2022_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(billi_2022_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(billi_2022_ctrl_human_fb, cluster_column='cluster_robust')
billi_2022_ctrl_human_fb.write_h5ad(billi_2022_dir + '/billi_2022_ctrl_human_fb_robust.h5')

### Boothby 2021

assign_cats(boothby_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.95, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(boothby_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

boothby_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(boothby_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(boothby_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(boothby_2021_ctrl_human_fb, cluster_column='cluster_robust')
boothby_2021_ctrl_human_fb.write_h5ad(boothby_2021_dir + '/boothby_2021_ctrl_human_fb_robust.h5')

### Burja 2022

assign_cats(burja_2022_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.99, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(burja_2022_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

burja_2022_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(burja_2022_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(burja_2022_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(burja_2022_ctrl_human_fb, cluster_column='cluster_robust')
burja_2022_ctrl_human_fb.write_h5ad(burja_2022_dir + '/burja_2022_ctrl_human_fb_robust.h5')

### Deng 2021

assign_cats(deng_2021_scar_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.95, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(deng_2021_scar_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

deng_2021_scar_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(deng_2021_scar_human_fb.obs['cluster_robust']))]
sc.pl.umap(deng_2021_scar_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(deng_2021_scar_human_fb, cluster_column='cluster_robust')
deng_2021_scar_human_fb.write_h5ad(deng_2021_dir + '/deng_2021_scar_human_fb_robust.h5')

### Gao 2021

assign_cats(gao_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.75, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(gao_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

gao_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(gao_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(gao_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(gao_2021_ctrl_human_fb, cluster_column='cluster_robust')
gao_2021_ctrl_human_fb.write_h5ad(gao_2021_dir + '/gao_2021_ctrl_human_fb_robust.h5')

### Gaydosik 2020

assign_cats(gaydosik_2020_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.45, quantile_gene_sel=0.8, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(gaydosik_2020_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

gaydosik_2020_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(gaydosik_2020_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(gaydosik_2020_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(gaydosik_2020_ctrl_human_fb, cluster_column='cluster_robust')
gaydosik_2020_ctrl_human_fb.write_h5ad(gaydosik_2020_dir + '/gaydosik_2020_ctrl_human_fb_robust.h5')

### Gur 2022

assign_cats(gur_2022_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.9, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(gur_2022_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

gur_2022_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(gur_2022_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(gur_2022_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(gur_2022_ctrl_human_fb, cluster_column='cluster_robust')
gur_2022_ctrl_human_fb.write_h5ad(gur_2022_dir + '/gur_2022_ctrl_human_fb_robust.h5')

### He 2020

assign_cats(he_2020_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.9, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(he_2020_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

he_2020_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(he_2020_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(he_2020_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(he_2020_ctrl_human_fb, cluster_column='cluster_robust')
he_2020_ctrl_human_fb.write_h5ad(he_2020_dir + '/he_2020_ctrl_human_fb_robust.h5')

### Hughes 2020

assign_cats(hughes_2020_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.75, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(hughes_2020_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)
hughes_2020_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(hughes_2020_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(hughes_2020_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(hughes_2020_ctrl_human_fb, cluster_column='cluster_robust')
hughes_2020_ctrl_human_fb.write_h5ad(hughes_2020_dir + '/hughes_2020_ctrl_human_fb_robust.h5')

### Kim 2020

assign_cats(kim_2020_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.45, quantile_gene_sel=0.99, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(kim_2020_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

kim_2020_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(kim_2020_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(kim_2020_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(kim_2020_ctrl_human_fb, cluster_column='cluster_robust')
kim_2020_ctrl_human_fb.write_h5ad(kim_2020_dir + '/kim_2020_ctrl_human_fb_robust.h5')

### Liu 2021

assign_cats(liu_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.95, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(liu_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

liu_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(liu_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(liu_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(liu_2021_ctrl_human_fb, cluster_column='cluster_robust')
liu_2021_ctrl_human_fb.write_h5ad(liu_2021_dir + '/liu_2021_ctrl_human_fb_robust.h5')

### Mariottoni 2021

assign_cats(mariottoni_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.9, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(mariottoni_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

mariottoni_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(mariottoni_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(mariottoni_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(mariottoni_2021_ctrl_human_fb, cluster_column='cluster_robust')
mariottoni_2021_ctrl_human_fb.write_h5ad(mariottoni_2021_dir + '/mariottoni_2021_ctrl_human_fb_robust.h5')

### Mirizio 2020

assign_cats(mirizio_2020_scl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.4, quantile_gene_sel=0.95, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(mirizio_2020_scl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

mirizio_2020_scl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(mirizio_2020_scl_human_fb.obs['cluster_robust']))]
sc.pl.umap(mirizio_2020_scl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(mirizio_2020_scl_human_fb, cluster_column='cluster_robust')
mirizio_2020_scl_human_fb.write_h5ad(mirizio_2020_dir + '/mirizio_2020_scl_human_fb_robust.h5')

### Reynolds 2021

assign_cats(reynolds_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.95, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(reynolds_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', quantile_gene_sel=0.85,
            intermediate_states=True, diff=0.15, others_name='U', verbose=False)

reynolds_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(reynolds_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(reynolds_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(reynolds_2021_ctrl_human_fb, cluster_column='cluster_robust')
reynolds_2021_ctrl_human_fb.write_h5ad(reynolds_2021_dir + '/reynolds_2021_ctrl_human_fb_robust.h5')

### Rindler 2021

assign_cats(rindler_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.45, quantile_gene_sel=0.9, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(rindler_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

rindler_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(rindler_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(rindler_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(rindler_2021_ctrl_human_fb, cluster_column='cluster_robust')
rindler_2021_ctrl_human_fb.write_h5ad(rindler_2021_dir + '/rindler_2021_ctrl_human_fb_robust.h5')

### Solé-Boldo 2021

assign_cats(sole_2020_young_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.9, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(sole_2020_young_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

sole_2020_young_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(sole_2020_young_human_fb.obs['cluster_robust']))]
sc.pl.umap(sole_2020_young_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(sole_2020_young_human_fb, cluster_column='cluster_robust')
sole_2020_young_human_fb.write_h5ad(sole_2020_dir + '/sole_2020_young_human_fb_robust.h5')

### Tabib 2018

assign_cats(tabib_2018_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.85, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(tabib_2018_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

tabib_2018_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(tabib_2018_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(tabib_2018_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(tabib_2018_ctrl_human_fb, cluster_column='cluster_robust')
tabib_2018_ctrl_human_fb.write_h5ad(tabib_2018_dir + '/tabib_2018_ctrl_human_fb_robust.h5')

### Tabib 2021

assign_cats(tabib_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.4, quantile_gene_sel=0.75, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(tabib_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

tabib_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(tabib_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(tabib_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(tabib_2021_ctrl_human_fb, cluster_column='cluster_robust')
tabib_2021_ctrl_human_fb.write_h5ad(tabib_2021_dir + '/tabib_2021_ctrl_human_fb_robust.h5')

### Tabula Sapiens 2021

assign_cats(tsc_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.4, quantile_gene_sel=0.75, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(tsc_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

tsc_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(tsc_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(tsc_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(tsc_2021_ctrl_human_fb, cluster_column='cluster_robust')
tsc_2021_ctrl_human_fb.write_h5ad(tabib_2021_dir + '/tsc_2021_ctrl_human_fb_robust.h5')

### Theocarditis 2020

assign_cats(theo_2020_ctrl_human_dm_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.9, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(theo_2020_ctrl_human_dm_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

theo_2020_ctrl_human_dm_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(theo_2020_ctrl_human_dm_fb.obs['cluster_robust']))]
sc.pl.umap(theo_2020_ctrl_human_dm_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(theo_2020_ctrl_human_dm_fb, cluster_column='cluster_robust')
theo_2020_ctrl_human_dm_fb.write_h5ad(theo_2020_dir + '/theo_2020_ctrl_human_dm_fb_robust.h5')

### Theocarditis 2021

assign_cats(theo_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.95, 
            key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(theo_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4,  key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

theo_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(theo_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(theo_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(theo_2021_ctrl_human_fb, cluster_column='cluster_robust')
theo_2021_ctrl_human_fb.write_h5ad(theo_2021_dir + '/theo_2021_ctrl_human_fb_robust.h5')

### TSC 2021

assign_cats(tsc_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.95, 
            key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(tsc_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4,  key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

tsc_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(tsc_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(tsc_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(tsc_2021_ctrl_human_fb, cluster_column='cluster_robust')
tsc_2021_ctrl_human_fb.write_h5ad(tsc_2021_dir + '/tsc_2021_ctrl_human_fb_robust.h5')

### Vorstandlechner 2020

assign_cats(vors_2020_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.40, quantile_gene_sel=0.9, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(vors_2020_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

vors_2020_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(vors_2020_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(vors_2020_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(vors_2020_ctrl_human_fb, cluster_column='cluster_robust')
vors_2020_ctrl_human_fb.write_h5ad(vors_2020_dir + '/vors_2020_ctrl_human_fb_robust.h5')

### Vorstandlechner 2021

assign_cats(vors_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.5, quantile_gene_sel=0.95, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(vors_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

vors_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(vors_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(vors_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(vors_2021_ctrl_human_fb, cluster_column='cluster_robust')
vors_2021_ctrl_human_fb.write_h5ad(vors_2021_dir + '/vors_2021_ctrl_human_fb_robust.h5')

### Xu 2021

assign_cats(xu_2021_ctrl_human_fb, dict_cats=dict_cats_clusters_robust, min_score=0.45, quantile_gene_sel=0.9, key_added='cluster_robust', others_name='U', verbose=False)
assign_cats(xu_2021_ctrl_human_fb, column_groupby='cluster_robust', dict_cats=dict_cats_axes_robust, min_score=0.4, key_added='axis_robust', 
            quantile_gene_sel=0.95, intermediate_states=True, diff=0.05, others_name='U', verbose=False)

xu_2021_ctrl_human_fb.uns['cluster_robust_colors'] = [dict_colors_human[i] if i in dict_colors_human else '#bcbcbc' for i in sorted(set(xu_2021_ctrl_human_fb.obs['cluster_robust']))]
sc.pl.umap(xu_2021_ctrl_human_fb, color=['Internal sample identifier', 'leiden', 'axis_robust', 'cluster_robust'], legend_loc='on data', cmap=magma, use_raw=False, ncols=2)
plot_score_graph(xu_2021_ctrl_human_fb, cluster_column='cluster_robust')
xu_2021_ctrl_human_fb.write_h5ad(xu_2021_dir + '/xu_2021_ctrl_human_fb_robust.h5')

## Reevaluate the presence of clusters for each dataset

list_all_datasets_human = [ahlers_2022_young_human_fb, billi_2022_ctrl_human_fb, boothby_2021_ctrl_human_fb, burja_2022_ctrl_human_fb,
                     deng_2021_scar_human_fb, gao_2021_ctrl_human_fb, gaydosik_2020_ctrl_human_fb, gur_2022_ctrl_human_fb,
                     he_2020_ctrl_human_fb, hughes_2020_ctrl_human_fb, kim_2020_ctrl_human_fb, 
                     liu_2021_ctrl_human_fb, mariottoni_2021_ctrl_human_fb,
                     mirizio_2020_scl_human_fb, reynolds_2021_ctrl_human_fb, rindler_2021_ctrl_human_fb,
                     sole_2020_young_human_fb, tabib_2018_ctrl_human_fb, tabib_2021_ctrl_human_fb,
                     tsc_2021_ctrl_human_fb, theo_2020_ctrl_human_dm_fb, theo_2021_ctrl_human_fb, 
                     vors_2020_ctrl_human_fb, vors_2021_ctrl_human_fb, xu_2021_ctrl_human_fb]

list_names_human = [adata.obs['Author'].values[0] + ' ' + str(int(adata.obs['Year'].values[0])) for adata in list_all_datasets_human]

**We store these variables for following notebooks. This forces revisit this notebook in case we add a new dataset and also to make sure there are no dataset inconsistencies across notebooks.**

%store list_all_datasets_human
%store list_datasets_human

%store list_accepted_clusters_human
%store list_accepted_axis_human
%store list_names_human

plot_adata_cluster_properties(dict_cats_clusters=dict_cats_clusters_robust, list_datasets=list_all_datasets_human, what='presence', cluster_name='cluster_robust', axis_name='axis_robust')

plot_adata_cluster_properties(dict_cats_clusters=dict_cats_clusters_robust, list_datasets=list_all_datasets_human, what='percentage', cluster_name='cluster_robust', axis_name='axis_robust')

plot_adata_cluster_properties(dict_cats_clusters=dict_cats_clusters_robust, list_datasets=list_all_datasets_human, what='axis', cluster_name='cluster_robust', axis_name='axis_robust')

### Plotting basic dataset statistics

df_per = plot_adata_cluster_properties(dict_cats_clusters=dict_cats_clusters_robust, list_datasets=list_all_datasets_human, what='percentage', cluster_name='cluster_robust', axis_name='axis_robust')
df_per[df_per == 0] = np.NaN
df_per = df_per.astype(float)

df_per

fig, ax = plt.subplots(1, 1, figsize=(8,4.5))
for cluster_idx, cluster in enumerate(list_accepted_clusters_human):
    ax.bar(cluster_idx, np.nanmean(df_per[cluster].values), yerr=np.nanstd(df_per[cluster].values), color=dict_colors_human[cluster])

ax.set_xticks(range(len(list_accepted_clusters_human)))
ax.set_xticklabels(list_accepted_clusters_human)
ax.spines.right.set_visible(False)
ax.spines.top.set_visible(False)
None

fig, ax = plt.subplots(1, 1, figsize=(8,4.5))
for row_idx, row in enumerate(df_per.index):
    bottom=0
    for cluster_idx, cluster in enumerate(list_accepted_clusters_human):
        value = float(df_per.loc[row, cluster])
        if not np.isnan(value):
            value /= np.nansum(df_per.loc[row].values)
            ax.bar(row_idx, value, bottom=bottom, color=dict_colors_human[cluster])
            bottom += value

ax.set_xticks(range(len(df_per.index)))
ax.set_xticklabels([' '.join(i.split('-')[:2]) for i in df_per.index], rotation=90)
ax.spines.right.set_visible(False)
ax.spines.top.set_visible(False)
None

## Plotting all Adatas

fig, axs = plt.subplots(5, 5, figsize=(5 * 4, 5 * 4))

for ax in axs.ravel()[len(list_all_datasets_human) - len(axs) :]:
    ax.set_axis_off()

for adata, name, idx in zip(list_all_datasets_human, list_names_human, range(len(list_all_datasets_human))):
    sc.pl.umap(adata, color=['cluster_robust'], legend_loc='on data', show=False, ax = axs.ravel()[idx], legend_fontsize=13,
               title=str(adata.obs['Author'].iloc[0]) + ' ' + str(int(adata.obs['Year'].iloc[0])) + ' human', size=16, cmap=magma, frameon=False)

## PAGA 

fig, axs = plt.subplots(5, 5, figsize=(5 * 4, 5 * 4))

for ax in axs.ravel():
    ax.set_axis_off()
    
for adata, name, idx in zip(list_all_datasets_human, list_names_human, range(len(list_all_datasets_human))):
    sc.tl.paga(adata, groups='cluster_robust')
    sc.pl.paga(adata, ax=axs.ravel()[idx], frameon=False, show=False, 
               title=str(adata.obs['Author'].iloc[0]) + ' ' + str(int(adata.obs['Year'].iloc[0])) + ' human')

fig, axs = plt.subplots(5, 5, figsize=(5 * 4, 5 * 4))

for ax in axs.ravel()[len(list_all_datasets_human) - len(axs) :]:
    ax.set_axis_off()
    
for adata, name, idx in zip(list_all_datasets_human, list_names_human, range(len(list_all_datasets_human))):
    sc.tl.paga(adata, groups='cluster_robust')
    sc.pl.paga(adata, ax=axs.ravel()[idx], frameon=False, show=False, solid_edges='connectivities_tree', 
               title=str(adata.obs['Author'].iloc[0]) + ' ' + str(int(adata.obs['Year'].iloc[0])) + ' human')

## Combined PAGA 

### Using connectivities_tree

df_all = pd.DataFrame(0, index=list_accepted_clusters_human + ['U'], columns=list_accepted_clusters_human + ['U'])

for adata in tqdm(list_all_datasets_human):
    dfsub = pd.DataFrame(adata.uns['paga']['connectivities_tree'].todense(), 
                  columns=adata.obs['cluster_robust'].cat.categories, 
                  index=adata.obs['cluster_robust'].cat.categories)
    
    df_all.loc[dfsub.index, dfsub.index] = df_all.loc[dfsub.index, dfsub.index] + dfsub
    
df_all = df_all.iloc[:-2, :-2] # To remove U and T1

df_all[df_all < 0.8] = 0
df_all = df_all ** 1.25

df_all = (df_all + df_all.transpose()) / 2

df_all

# CHOSEN SEED IS 0
for seed in range(10):
    fig, ax = plt.subplots(1, 1, figsize=(5, 4))
    G = nx.convert_matrix.from_pandas_adjacency(df_all)
    pos=nx.spring_layout(G, seed=seed)

    edges = G.edges()
    weights = [G[u][v]['weight'] for u,v in edges]

    nx.draw_networkx(G, pos, width=weights, node_color=[dict_colors_human[i] for i in df_all.index], ax=ax)
    plt.axis('off')
    print(seed)
    plt.show()

### Using connectivities

df_all = pd.DataFrame(0, index=list_accepted_clusters_human + ['U'], columns=list_accepted_clusters_human + ['U'])

for adata in list_all_datasets_human:
    dfsub = pd.DataFrame(adata.uns['paga']['connectivities'].todense(), 
                  columns=adata.obs['cluster_robust'].cat.categories, 
                  index=adata.obs['cluster_robust'].cat.categories)
    
    df_all.loc[dfsub.index, dfsub.index] = df_all.loc[dfsub.index, dfsub.index] + dfsub
    
df_all = df_all.iloc[:-1, :-1] # To remove U

df_all[df_all < 2.3] = 0

df_all = (df_all + df_all.transpose()) / 2

df_all = df_all

# CHOSEN SEED IS 3
for seed in range(10):
    fig, ax = plt.subplots(1, 1, figsize=(5, 4))
    G = nx.convert_matrix.from_pandas_adjacency(df_all)
    pos=nx.spring_layout(G, seed=seed)

    edges = G.edges()
    weights = [G[u][v]['weight'] for u,v in edges]

    nx.draw_networkx(G, pos, width=weights, node_color=[dict_colors_human[i] for i in df_all.index], ax=ax)
    plt.axis('off')
    print(seed)
    plt.show()

os.system("jupyter nbconvert --to html 3H_obtain_robust_markers.ipynb")
